requirements.txt
huggingface_hub==0.26.5
snowflake==1.0.2
snowflake_connector_python==3.12.3
snowflake_ml_python==1.7.2
snowflake_snowpark_python==1.25.0
nltk==3.9.1
pandas==2.2.3
python-dotenv==1.0.1
llama_index==0.12.2
fastapi==0.115.6
huggingface-hub==0.26.5
pydantic==2.9.2
python-dotenv==1.0.1
together==1.3.5
uvicorn==0.34.0
folium==0.19.4
geopy==2.4.1